Se cambia el de tipo de cantidad de Dinero, de int a Double. Lo mismo modifica los tests de test/fiuba/algo3/tp2/algopoly/model/CarcelTest.java test/fiuba/algo3/tp2/algopoly/PrimeraEntregaTest.java test/fiuba/algo3/tp2/algopoly/model/estado/EstadoTest.java. Y se modifican src/fiuba/algo3/tp2/algopoly/model/Dinero.java src/fiuba/algo3/tp2/algopoly/model/casillero/AvanceDinamico.java src/fiuba/algo3/tp2/algopoly/model/casillero/RetrocesoDinamico.java, los ultimos dos se castea a int la cantidad del getCapital()

+++ src/fiuba/algo3/tp2/algopoly/model/Dinero.java
@@ -2,13 +2,13 @@ package fiuba.algo3.tp2.algopoly.model;
 
 public class Dinero {
 
-    private int cantidad;
+    private double cantidad;
 
     public Dinero(int cantidad) {
         this.cantidad = cantidad;
     }
 
-    public int getCantidad() {  //TODO private
+    public double getCantidad() {  //TODO private
         return this.cantidad;
     }
